audit: remove 7 redundant type-conversion round-trips + add audit document
A project-wide audit of unnecessary type conversions, run downstream of
the LongArray-substrate decision. Two outputs: tier-1 simplifications
applied directly, tier-2 design questions documented in
docs/general/type-conversion-audit.md.

## Tier-1 fixes (this commit)

Pattern 1: `Long.toInt().toShort()` -> `Long.toShort()`
  Same bit pattern, one fewer narrowing op. Long.toShort() is defined
  to return the low 16 bits — exactly what the chain produces.

  src/commonMain/kotlin/io/github/kotlinmania/klang/bitwise/BitShiftEngine.kt:1410, 1436, 1462, 1500
  src/commonMain/kotlin/io/github/kotlinmania/klang/mem/PackedBuffer.kt:102

Pattern 2: `Short.toInt().toLong()` -> `Short.toLong()`
  Short.toLong() is the direct sign-extending widening. The middle
  Int step did sign-extend-to-32 followed by sign-extend-to-64, two
  ops where one suffices.

  src/commonMain/kotlin/io/github/kotlinmania/klang/internal/runtime/Runtime.kt:125, 126
  src/commonMain/kotlin/io/github/kotlinmania/klang/fp/CBF16.kt:62, 68

7 conversion ops removed. Identical semantics. Tests green
(./gradlew macosArm64Test -> 439/0/0).

## Tier-2 documented (no code change yet)

Two API-shape questions captured in
docs/general/type-conversion-audit.md as proposals for separate PRs:

  A. PackedBuffer.setShort(addr, value: Short) probably should take
     Int to match setByte(addr, value: Int). Currently every 16-bit
     store does Int -> Short -> Int through-and-back via the type
     system. Affects ~20 call sites in F16Math, Float128Math,
     BitShiftEngine.

  B. BitShiftEngine could expose Int-overload shifts (leftShift32,
     rightShift32, unsignedRightShift32) so 32-bit operands don't
     pay Int->Long widen and Long->Int narrow on every call. Affects
     ~30 call sites in CBF16, HexShift, BitwiseOps, BitPrimitives,
     Float32Math.

Both deferred for benchmark-backed evaluation.

## Tier-3 documented (intentional patterns, not bugs)

Three patterns captured to forestall future 'cleanup' that would
remove them by mistake:

  - .toLong() at engine call sites (the engine API requires Long)
  - (bits and 0xFF).toUByte() in CE8M0/CUE4M3 (documents the
    byte-width assumption)
  - HeapUInt128.toIntArray() in Float128Math (intentional snapshot
    for read-modify-write)

## docs/general/type-conversion-audit.md (new)

Structured report with the three tiers, file-and-line table for tier-1
fixes, expanded rationale for tier-2 proposals, the grep recipes used
to find these patterns (so the audit is reproducible), and
cross-references to why-longarray-storage.md and philosophy-and-design.md.

## Verification

  ./gradlew compileKotlinMacosArm64 macosArm64Test  -> green (439/0/0)

+++ src/commonMain/kotlin/io/github/kotlinmania/klang/fp/CBF16.kt
@@ -59,13 +59,17 @@ class CBF16 private constructor(private val bits: Short) {
 
     fun isNaN(): Boolean {
         val engine = BitShiftEngine(BitShiftConfig.defaultMode, 32)
-        val e = (engine.unsignedRightShift(bits.toInt().toLong(), 7).value.toInt() and 0xFF)
+        // Short.toLong() is direct sign-extending widening; round-tripping
+        // through .toInt().toLong() did the same job in two operations.
+        val e = (engine.unsignedRightShift(bits.toLong(), 7).value.toInt() and 0xFF)
         val f = bits.toInt() and 0x7F
         return e == 0xFF && f != 0
     }
     fun isInf(): Boolean {
         val engine = BitShiftEngine(BitShiftConfig.defaultMode, 32)
-        val e = (engine.unsignedRightShift(bits.toInt().toLong(), 7).value.toInt() and 0xFF)
+        // Short.toLong() is direct sign-extending widening; round-tripping
+        // through .toInt().toLong() did the same job in two operations.
+        val e = (engine.unsignedRightShift(bits.toLong(), 7).value.toInt() and 0xFF)
         val f = bits.toInt() and 0x7F
         return e == 0xFF && f == 0
     }
